7.12 - Add String to LogAnalyzer Constructor
Modify the LogANalyzer class so that it has a constructor that can take the name of the log file to be analyzed. Have this constructor pass the file name to the constructor of the LogFileReader class. Use the LogfileCreator class to create your own file of random log entries, and analyze the data.